feat(CL): continue strategy pattern to simplify swaps
