+++ x/concentrated-liquidity/internal/math/math.go
@@ -4,28 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-type swapStrategy interface {
-	GetNextSqrtPriceFromInput(sqrtPriceCurrent, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext sdk.Dec)
-	ComputeSwapStep(sqrtPriceCurrent, sqrtPriceTarget, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext, amountIn, amountOut sdk.Dec)
-	SetLiquidityDeltaSign(sdk.Dec) sdk.Dec
-	SetNextTick(int64) sdk.Int
-}
-
-type zeroForOneStrategy struct{}
-
-var _ swapStrategy = (*zeroForOneStrategy)(nil)
-
-type oneForZeroStrategy struct{}
-
-var _ swapStrategy = (*oneForZeroStrategy)(nil)
-
-func NewSwapStrategy(zeroForOne bool) swapStrategy {
-	if zeroForOne {
-		return &zeroForOneStrategy{}
-	}
-	return &oneForZeroStrategy{}
-}
-
 // liquidity0 takes an amount of asset0 in the pool as well as the sqrtpCur and the nextPrice
 // sqrtPriceA is the smaller of sqrtpCur and the nextPrice
 // sqrtPriceB is the larger of sqrtpCur and the nextPrice
@@ -94,58 +72,6 @@ func CalcAmount1Delta(liq, sqrtPriceA, sqrtPriceB sdk.Dec, roundUp bool) sdk.Dec
 		return liq.Mul(diff).Ceil()
 	}
 	return liq.Mul(diff)
-}
-
-// ComputeSwapStep calculates the amountIn, amountOut, and the next sqrtPrice given current price, price target, tick liquidity, and amount available to swap
-// lte is reference to "less than or equal", which determines if we are moving left or right of the current price to find the next initialized tick with liquidity
-func (s *zeroForOneStrategy) ComputeSwapStep(sqrtPriceCurrent, sqrtPriceTarget, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext, amountIn, amountOut sdk.Dec) {
-	amountIn = CalcAmount0Delta(liquidity, sqrtPriceTarget, sqrtPriceCurrent, false)
-	if amountRemaining.GTE(amountIn) {
-		sqrtPriceNext = sqrtPriceTarget
-	} else {
-		sqrtPriceNext = s.GetNextSqrtPriceFromInput(sqrtPriceCurrent, liquidity, amountRemaining)
-	}
-	amountIn = CalcAmount0Delta(liquidity, sqrtPriceNext, sqrtPriceCurrent, false)
-	amountOut = CalcAmount1Delta(liquidity, sqrtPriceNext, sqrtPriceCurrent, false)
-
-	return sqrtPriceNext, amountIn, amountOut
-}
-
-func (s *oneForZeroStrategy) ComputeSwapStep(sqrtPriceCurrent, sqrtPriceTarget, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext, amountIn, amountOut sdk.Dec) {
-	amountIn = CalcAmount1Delta(liquidity, sqrtPriceTarget, sqrtPriceCurrent, false)
-	if amountRemaining.GTE(amountIn) {
-		sqrtPriceNext = sqrtPriceTarget
-	} else {
-		sqrtPriceNext = s.GetNextSqrtPriceFromInput(sqrtPriceCurrent, liquidity, amountRemaining)
-	}
-	amountIn = CalcAmount1Delta(liquidity, sqrtPriceNext, sqrtPriceCurrent, false)
-	amountOut = CalcAmount0Delta(liquidity, sqrtPriceNext, sqrtPriceCurrent, false)
-
-	return sqrtPriceNext, amountIn, amountOut
-}
-
-func (s *zeroForOneStrategy) GetNextSqrtPriceFromInput(sqrtPriceCurrent, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext sdk.Dec) {
-	return GetNextSqrtPriceFromAmount0RoundingUp(sqrtPriceCurrent, liquidity, amountRemaining)
-}
-
-func (s *oneForZeroStrategy) GetNextSqrtPriceFromInput(sqrtPriceCurrent, liquidity, amountRemaining sdk.Dec) (sqrtPriceNext sdk.Dec) {
-	return GetNextSqrtPriceFromAmount1RoundingDown(sqrtPriceCurrent, liquidity, amountRemaining)
-}
-
-func (s *zeroForOneStrategy) SetLiquidityDeltaSign(deltaLiquidity sdk.Dec) sdk.Dec {
-	return deltaLiquidity.Neg()
-}
-
-func (s *oneForZeroStrategy) SetLiquidityDeltaSign(deltaLiquidity sdk.Dec) sdk.Dec {
-	return deltaLiquidity
-}
-
-func (s *zeroForOneStrategy) SetNextTick(nextTick int64) sdk.Int {
-	return sdk.NewInt(nextTick - 1)
-}
-
-func (s *oneForZeroStrategy) SetNextTick(nextTick int64) sdk.Int {
-	return sdk.NewInt(nextTick)
 }
 
 // getNextSqrtPriceFromAmount0RoundingUp utilizes the current squareRootPrice, liquidity of denom0, and amount of denom0 that still needs
